Updated signup route to align with API docs
